+++ core/src/PluginHook.ts
@@ -1,6 +1,6 @@
 import { PluginComponent } from './PluginComponent';
 import { Emitter } from './EventEmitter';
-import { MiddlewareCollection, DefaultEvents, Events } from './utils/Interfaces';
+import { MiddlewareCollection, DefaultEvents, Events } from './interfaces';
 
 export abstract class PluginHook<T extends Events = DefaultEvents> extends PluginComponent {
   protected middlewareCollection!: MiddlewareCollection<T | DefaultEvents>;
