+++ core/src/constants.ts
