+++ core/src/index.ts
@@ -1,6 +1,6 @@
 import prompt from 'prompts';
 import { config } from 'dotenv';
-import { ProjectConfigFile } from './utils/Interfaces';
+import { ProjectConfigFile } from './interfaces';
 
 // Load .env variables into process.env
 config();
@@ -12,7 +12,6 @@ export * from './Logger';
 export * from './JovoCli';
 export * from './EventEmitter';
 export * from './PluginHook';
-export * from './utils';
 export * from './PluginCommand';
 export * from './Task';
 export * from './Project';
@@ -20,6 +19,11 @@ export * from './JovoCliPlugin';
 export * from './JovoCliError';
 export * from './JovoUserConfig';
 export * from './Config';
+export * from './utilities';
+export * from './interfaces';
+export * from './validators';
+export * from './prompts';
+export * from './constants';
 
 export class ProjectConfig {
   constructor(config: ProjectConfigFile) {
