+++ core/src/spinnies.d.ts
